Revert "Done Bang "
