--- app/src/main/res/drawable/background_round_white.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#FAF5F1" />
-    <corners android:radius="12dp" />
-</shape>
