--- app/src/main/res/drawable/rounded_background_c4b7af.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <solid android:color="#C4B7AF" />
-    <corners android:radius="8dp" />
-    <stroke android:color="@color/black" android:width="1dp"/> </shape>
