--- app/src/main/res/drawable/rounded_background_faf5f1.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <stroke android:width="1dp" android:color="@color/black" />
-    <solid android:color="#FAF5F1" />
-    <corners android:radius="5dp" />
-</shape>
