--- app/src/main/res/drawable/rounded_button_8f7a6e.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#8F7A6E" />
-    <corners android:radius="8dp" />
-</shape>
