--- app/src/main/res/drawable/rounded_button_a41f13.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#A41F13" />
-    <corners android:radius="8dp" />
-</shape>
