--- app/src/main/res/drawable/rounded_dialog_background.xml
@@ -1,8 +1,0 @@
-<!-- res/drawable/rounded_dialog_background.xml -->
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="@android:color/white"/>
-    <corners android:radius="12dp"/>
-    <stroke android:color="#DDDDDD" android:width="1dp"/>
-    <!-- Opsional: Tambahkan shadow jika Anda mau, tapi ini lebih kompleks di XML -->
-</shape>
