--- app/src/main/res/drawable/rounded_edittext_white.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <solid android:color="#FFFFFF" />
-    <corners android:radius="8dp" />
-    <stroke android:color="#DDDDDD" android:width="1dp"/>
-</shape>
