+++ app/src/main/res/layout/fragment_vehicle_park.xml
@@ -4,8 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".VehicleParkFragment"
-    android:background="@color/white">
+    tools:context=".VehicleParkFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
