+++ app/src/main/res/layout/report_tagihan_fragment.xml
@@ -160,9 +160,8 @@
                     <TextView
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:fontFamily="@font/cabin"
                         android:text="No Virtual Account"
-                        android:textStyle="bold"
+                        android:fontFamily="@font/cabin"
                         android:textColor="#53141D"
                         android:textSize="16sp" />
 
@@ -211,12 +210,12 @@
             android:textSize="26sp"/>
 
         <!-- Tombol Refresh -->
-        <Button
-            android:id="@+id/refreshButton"
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/backButton"
             android:layout_width="match_parent"
             android:layout_height="54dp"
             android:text="Refresh"
-            android:backgroundTint="@color/maroon"
+            android:background="@drawable/button_style"
             android:textColor="#FFFFFF"
             android:layout_marginTop="35dp"
             android:fontFamily="@font/cabin"
