+++ app/src/main/res/values/themes.xml
@@ -6,13 +6,4 @@
     </style>
 
     <style name="Theme.KosanKu" parent="Base.Theme.KosanKu" />
-
-    <style name="CustomDialogTheme" parent="Theme.AppCompat.Light.Dialog">
-        <item name="android:windowBackground">@android:color/transparent</item>
-        <item name="android:windowIsFloating">true</item>
-        <item name="android:windowNoTitle">true</item>
-        <!-- Warna overlay abu-abu transparan -->
-        <item name="android:backgroundDimEnabled">true</item>
-        <item name="android:backgroundDimAmount">0.6</item> <!-- 0.0 (transparent) to 1.0 (fully opaque) -->
-    </style>
 </resources>
